Fix MySQL Error 1093 for Delete_Where_optional_navigation_predicate test

+++ test/EFCore.MySql.FunctionalTests/BulkUpdates/NorthwindBulkUpdatesMySqlTest.cs
@@ -735,8 +735,11 @@ WHERE EXISTS (
 
     public override async Task Delete_Where_optional_navigation_predicate(bool async)
     {
-        await base.Delete_Where_optional_navigation_predicate(async);
-        AssertSql(
+        if (AppConfig.ServerVersion.Supports.DeleteWithSelfReferencingSubquery)
+        {
+            await base.Delete_Where_optional_navigation_predicate(async);
+
+            AssertSql(
 """
 DELETE `o`
 FROM `Order Details` AS `o`
@@ -747,6 +750,14 @@ WHERE EXISTS (
     LEFT JOIN `Customers` AS `c` ON `o1`.`CustomerID` = `c`.`CustomerID`
     WHERE (`c`.`City` LIKE 'Se%') AND ((`o0`.`OrderID` = `o`.`OrderID`) AND (`o0`.`ProductID` = `o`.`ProductID`)))
 """);
+        }
+        else
+        {
+            // Not supported by MySQL and older MariaDB versions:
+            //     Error Code: 1093. You can't specify target table 'o' for update in FROM clause
+            await Assert.ThrowsAsync<MySqlException>(
+                () => base.Delete_Where_optional_navigation_predicate(async));
+        }
     }
 
     public override async Task Delete_with_join(bool async)
